--- flaskr/templates/auth/login.html
@@ -1,9 +1,0 @@
-{% extends 'base.html' %}
-
-{% block header %}
-  <h1>{% block title %}Log In{% endblock %}</h1>
-{% endblock %}
-
-{% block content %}
-
-{% endblock %}
