--- flaskr/templates/blog/index.html
@@ -1,17 +1,0 @@
-{% extends 'base.html' %}
-{% block header %}
-  <h1>{% block title %}Tracks{% endblock %}</h1>
-{% endblock %}
-{% block content %}
-  {% for post in posts %}
-    <article class="post">
-      <header>
-          <h1>{{ post['id'], post['title'], post['artist'], post['genre'] }}</h1>
-      </header>
-      <p class="body">{{ post['body'] }}</p>
-    </article>
-    {% if not loop.last %}
-      <hr>
-    {% endif %}
-  {% endfor %}
-{% endblock %}
