+++ tracks/data.sql
@@ -1,0 +1,7 @@
+CREATE TABLE tracks (
+  id INTEGER PRIMARY KEY AUTOINCREMENT,
+  title TEXT UNIQUE NOT NULL,
+  artist TEXT NOT NULL,
+  genre TEXT NOT NULL,
+  track_length INTEGER NOT NULL
+);
