+++ tracks/db.py
@@ -1,0 +1,54 @@
+import sqlite3
+
+import click
+from flask import current_app, g
+
+
+def get_db():
+    g.db = sqlite3.connect(
+        current_app.config['DATABASE'],
+        detect_types=sqlite3.PARSE_DECLTYPES
+    )
+    g.db.row_factory = sqlite3.Row
+    return g.db
+
+
+def close_db(e=None):
+    db = g.pop('db', None)
+
+    if db is not None:
+        db.close()
+
+
+def init_db():
+    db = get_db()
+
+    with current_app.open_resource('data.sql') as f:
+        db.executescript(f.read().decode('utf8'))
+
+
+def fill_db():
+    db = get_db()
+
+    with current_app.open_resource('data_add.sql') as f:
+        db.executescript(f.read().decode('utf8'))
+
+
+@click.command('init-db')
+def init_db_command():
+    """Clear the existing data and create new tables."""
+    init_db()
+    click.echo('Initialized the database.')
+
+
+@click.command('fill-db')
+def fill_db_command():
+    """Clear the existing data and create new tables."""
+    fill_db()
+    click.echo('Fill the database.')
+
+
+def init_app(app):
+    app.teardown_appcontext(close_db)
+    app.cli.add_command(init_db_command)
+    app.cli.add_command(fill_db_command)
