+++ tracks/templates/blog/cnt_genre.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+{% block title %}
+    <h2> Quantity of tracks in genre </h2>
+{% endblock %}
+{% block head %}
+{% endblock %}
+{% block content %}
+    <h1>All unique artist:</h1>
+    <p>{{ cg }}</p>
+{% endblock %}
