+++ tracks/templates/blog/main.html
@@ -1,0 +1,16 @@
+{% extends "base.html" %}
+{% block title %}
+    <h2> Main </h2>>
+{% endblock %}
+{% block head %}
+{% endblock %}
+{% block content %}
+    <h1>Main page</h1>
+    <p>
+        <a href="{{ url_for('blog.names') }}">All unique artist</a><br>
+        <a href="{{ url_for('blog.tracks') }}">Quantity of tracks</a><br>
+        <a href="{{ url_for('blog.cnt_genre') }}">Quantity of tracks in genre</a><br>
+        <a href="{{ url_for('blog.cnt_genre') }}">Track names and there duration</a><br>
+        <a href="{{ url_for('blog.cnt_genre') }}">Average track duration</a><br>
+    </p>
+{% endblock %}
