+++ tracks/templates/blog/tracks.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+{% block title %}
+    <h2> Tracks </h2>
+{% endblock %}
+{% block head %}
+{% endblock %}
+{% block content %}
+    <h1>All unique artist:</h1>
+    <p>{{ tr }}</p>
+{% endblock %}
