+++ tracks/templates/blog/tracks_sec.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+{% block title %}
+    <h2> Track names and there duration </h2>
+{% endblock %}
+{% block head %}
+{% endblock %}
+{% block content %}
+    <h1>All unique artist:</h1>
+    <p>{{ ts }}</p>
+{% endblock %}
